Project Presentation
This is NO WHERE near complete. I need my team before I can work on it more (feel free to add to it as you see fit). Also, if you have the picture of the exercise we did in class today, that would be a great thing to add on the (current) last slide.